chore: Use ttsc to fix src/... imports in emitted JS.
Even though `tsc` will use the tsconfig.json `paths` key to resolve
modules during compilation path, it doesn't actually rewrite the
import to the resolved location.

So consumers still end up seeing `import { Css } from "src"` in say
`Icon.js` instead of `import { Css } from "../index"`.

ttsc is a tsc wrapper that adds exposes the tsc plugin api, and
we use that to add the typescript-transform-paths plugin to do this
rewriting for us.

This is the same setup we have in graphql-service so that its JS
has non-src imports in the emitted/production JS.