fix: Add 'position: relative;' to elements that wrap <VisuallyHidden />
The VisuallyHidden component creates an absolutely position element of height and width of 1px. This absolutely positioned element can act up when a parent element uses `display: grid`, and can cause the <VisuallyHidden /> element to take up space in the DOM when it shouldn't.
Updating to add 'position: relative;` keeps the VisuallyHidden element within the field's container and prevents it from being placed elsewhere in the app.