feat: ToggleChips
Expected use for ToggleChips will be used in the Filter's modal. Design had named them FilterChips, however we are naming them ambiguous to the location they were originally designed for in case we want to reuse them elsewhere.

Added an option to have the label for a field be hidden.