fix: Require 'label' in all Text/Number fields.
A label is required for accessibility purposes. We are not using the <VisuallyHidden /> for this input, as the label can be placed on the input element directly via aria-label. This allows us to omit possible styling issues that <VisuallyHidden /> has caused in the past due to it being absolutely positioned.