fix: SelectField interaction fixes (#226)
* fix: SelectField interaction fixes

Fix 1 - Handling focus state is now done without the use of React-Aria. It wasn't needed to use the focus ring, as the focus is displayed regardless of how the user interacts with it. (FocusRing is specifically for displaying keyboard focus). Instead we handle setting a focus value locally via setState
Fix 2 - Reset the SelectField/MultiSelectField input's value onBlur if not properly set.

* Adding tests to verify resetting the select field's input value on blur