fix: MultiSelectField to empty input value upon selecting an option
Usability issue discovered in user testing - When a user is actively using the MultiSelectField and chooses a single option, that option's text is used to populate the input. However, user expectations were for that input value to be removed and allow them to start typing to filter again without having to delete the current input.

Fix is to empty out the input value upon selection change for MultiSelectFields. If only one option is selected, then upon blurring the field the selected option should show in the input.