feat: Refactor filters using oo.
Sometimes FP and ADTs are great, but sometimes the "bring your own
behavior" of OO is great too.

This de-kinds the FilterDefs and instead makes the filters a list
of instances/classes that know how to render themselves.

This removes a lot of the type-inference gunk from FilterDefs where
it was trying to guess what kinds should be supported based on the
type of the filter.

Really we don't care what kind the filter is, as long as it can
handle the given filter value.