fix: Do not allow ListBox to show on top of trigger element (#299)
A bug was originally discovered when triggering the ListBox on top of the SelectField where an option would be immediately selected due to selection happening onPointerUp and the ListBox being triggered onPointerDown. This was originally fixed by setting 'shouldSelectOnPressUp: false' for the Options within the ListBox. Though, this traded one bug for another. Recently discovered was that the subsequent onPointerDown event fired after selecting an Option could have adverse effects, such as trigger a modal to close if the ListBox was rendered on top of the Modal's underlay element.

To address this issue, we went back to the root problem and now prevent the Menu from ever rendering above its trigger element. This is done by utilizing the OverlayPosition props, which include a 'maxHeight' property we use to define the ListBox's height. This ensures it will only take up the space available between the trigger and the edge of the viewport. A height is needed to be set due to the virtualization library used to render the ListBox's options.