fix: Update spacing and scrolling styles after QA with design
Margin between tabs and tab content should be 24px.

Adding in `overflow: overlay` even though it is only supported in Webkit browsers, though that is the majority of our users. Keeping `overflow: auto` defined as a fallback for any browser that do not support overflow: overlay. 'overlay' will prevent the scrollbars from pushing content in when are needed, which pushes the content inwards.