fix: Avoid unnecessary whitespace (#331)
Tab panel element is always rendered, which eliminates the possibility of TabContent benefiting from collapsable margins. This resulted in the margin-top: 24px being applied under an empty Tab Panel element, and causing extra white space. Now if the tabs are hidden, then the TabContent will not apply a margin-top definition.