feat: Update DatePicker to utilize TextFieldBase and Popover
TextFieldBase now supports the properties "endAdornment" and "inlineLabel". The "inlineLabel" will be used as part of the DateFilterField, and the "endAndornment" is to place the Calendar icon button at the end of the element.
The inlineLabel and endAdornment are only applied to the single line text field and ignored if multiline.