feat: Support TextAreaField styling for SchedulesV2 (#351)
* feat: Support TextAreaField styling for SchedulesV2

Changes current 'borderless' property to 'compound’.
Compound signifies that the field will be pair with another component and the wrapper of these components will handle supplying border styles. In these cases we’ll remove the extra 2px of height in TextFieldBase as the wrapping element is expected to add that border back in.
The “borderless” property changes the border color to be transparent and will keep the field height the same. Also a borderless field uses box-shadow for the focus styles.

Adds `preventNewLines` for TextAreaField to stop user from typing a new line and also replaces new line characters with an empty space if pasted from another source. `preventNewLines` is also a signal to the TextArea that we do not need a minimum height value.

* PR Feedback: Creates a TextFieldInternalProps interface that allows for Beam only customizations to our fields. Right now this only includes 'compound' as the fields have specific patterns they need to follow when they are put together in a Compound field. Also updates the GridTable Presentation Context to set 'compact'.