fix: RichTextField to initialize based on trix supplied init event (#365)
* fix: RichTextField to initialize based on trix supplied init event

Removing the document.getElementById call to grab the editor element and replacing with a useRef. This was safe as when using the RichTextField within a Portal it is possible that the element for the editor is not yet in the DOM.
Now listening to the 'trix-initialize' event in order to set the editor ref and call 'loadHTML'. This is not immediately available off of the element's ref because the 'editor' is attached to the 'trix-editor' web component using 'requestAnimationFrame', which waits for the event loop to clear before running its execution (similar to a setTimeout(..., 0))

* Attach event listener for 'trix-init' event to window. Use a unique id for targeting the component instance when the event fires

* PR: Further consolidation of logic.

* PR Feedback: Add comment