fix: ChipSelectField and ListBox updates (#394)
* fix: ChipSelectField and ListBox updates

ListBox now supports two sections, (1) Options, and (2) Persistent Actions. The Options section will scroll independently of the Persistent Actions, make the Actions always visible.
Sections will only be rendered if necessary, otherwise it will just be a flat list of options as it works today. If the onCreateNew prop is passed to ChipSelectField, then we'll create the two sections. Right now there are a lot of assumptions in the code that we'll only ever have at most two sections, and that the second section is the Persistent Actions, and there will only ever be a single persistent action. Prototype, period. :)

Bugs fixed:
- Auto scroll to focused item ONLY if using virutal focus (i.e. in SelectField where we use a Combobox and the browser focus stays within the input field). This fixes an issue where the ListBox was scrolling when trying to click on an option for the ChipSelectField. This was due to the focus moving as the user clicks and our auto-scroll repositioning the options. We now handle scrolling Virtuoso in the Option component itself if not using a virutal focus.
- Use 'offset' prop in 'useOverlayPosition' to properly position the ListBox. Our previous way of doing this would create issues when the menu opened above the trigger.
- Call 'onBlur' after user removes ChipTextField, regardless of whether they created a new option or not. If a new option was created, this should trigger formState's auto save behavior. Adds tests to ensure this is called with expected values.

Updates:
- Max height of Listbox has been set to 512px, per design request.

* PR feedback: add comments