fix: Allow Tooltip children to be ReactNode, which includes strings
The `children` prop was previously ReactElement was was require because we used 'cloneElement'. We no longer clone the element and can now allow for more types to be passed along such as strings.

Additionally adds the 'title' prop to the wrapping element when the tooltip is not showing and only if the title is of type string. This is to help with testing purposes and follows a similar pattern found from MUI.