fix: GridTable column 'fr' width values calculated to pixels (#462)
* fix: GridTable column 'fr' width values calculated to pixels

When the GridTable is first rendered then we use the existing width calculations for 'fr' units. Once the table is rendered, then the 'fr' units are calculated as pixel values. This is very similar to the logic in Blueprint's SchedulesV2. This properly allows for text-overflow to work and keeps all columns aligned, which was a recent issue with the new Beam "fixed" styling.
Uses a ResizeObserver for handling width recalculations in screensize changes.
Allows for having the resize observer listen to another element instead of the GridTable container (also required by SchedulesV2).
Also adds support for a minimum table width (as required by SchedulesV2).

* PR Feedback