fix: Avoid stack errors with cyclic rows.
We were unintentionally observing the RowState.rows field, which
was a ref to the rows themselves.

So when `RowState.rows.current` became "list of 100s of rows", we
were observizing them.

Which a) we should not do b/c it'd be a perf hint, but b) only blew
up if the rows had circular references which typically only happens
when our factories create dummy GraphQL fragment data.

So I reproduced that setup here, and then told mobx to not observe
the rows field.