fix: Do not 'restoreFocus' on Menu when it closes (#528)
The 'restoreFocus' causes React-Aria to try and retain focus on the element which triggered the Menu component. Though, if the menu closed in order to open a Modal, then we have conflicting signals to React Aria as to where focus should go - either in the Modal or back to the Menu's trigger. Removing 'restoreFocus' to no longer attempt to put the focus back to the Menu's button once closed. Will follow up with the Adobe team to provide this example and ask for guidance on how to go about this sort of behavior.