feat: Support setting group rows as initially collasped
Adds new property on GridDataRow, 'initCollapsed'. By setting this to true it will signal to the table to initially render this row as collapsed. As the name describes, this is only applied/read on initial render. If the user has also set up the 'persistCollapse' then any values available in the localStorage will trump those set on GridDataRow.

Includes tests