fix: Textfield 'errorInTooltip' to show the tooltip on an icon, not the whole field
The conditional wrapping of the tooltip when a textfield is in error causes some usability issues. Specifically, when the field is in error and the user then corrects the problem, it will rerender removing the conditional Tooltip wrapper. This rerender loses focus on the input, which can be very annoying for users.

Updating this error state to be inline with what is design for [Beam](https://www.figma.com/file/aWUE4pPeUTgrYZ4vaTYZQU/%E2%9C%A8Beam-Design-System?node-id=1944%3A14)