feat: SuperDrawerHeader (#734)
* feat: SuperDrawerHeader

Introduces the SuperDrawerHeader component. This component portals itself into the header of the SuperDrawer. Follows the existing pattern of our Modals.

This is a fairly large breaking change as we have completely changed how to define the header of the SuperDrawer. It removes all title/header configuration properties from the openInDrawer method.

* PR Feedback: Provide some default formatting for SuperDrawerHeader