feat: Allow parent rows to be selected independent of child row selections (#779)
When adding `GridDataRow.inferSelectedState: false` to a row, then it will no longer derived its selected state based on the state of its children rows. Instead the row must be selected specifically (or via its own parent) in order for it to be considered selected. This also removes the ability for the row to show in the 'indeterminate' state