fix: ScrollableContent bgColor bug fix
Previously setting the bgColor on ScrollableContent was only displaying the background color at 'height: 100%', which applied to only what was in view on the screen. Once the user scrolled, then you could see the background was not under all of the content. Switching to use flex-box and a flexGrow: 1 to have the ScrollableContent's element to extend the full height of the content within.

Additionally removing the bottom padding when using the fallback when ScrollableContent is not used. This is because of issues it causes with setting background colors elsewhere (like on tabs), but then the background is no applied to the ScrollableParent, so the extra padding would not include a background color, and would look weird. It would also be something not easily addressed by the consuming app, so just removing it in favor of requesting implementations to use ScrollableContent.