fix: OverlayTrigger to pass onClick value
It was discovered that on iOS that the OverlayTrigger was not invoking the menu to open. This appeared to be due to passing 'noop' to the trigger element's onClick property. For whatever reason on non-mobile browsers this appeared to still work find. The 'menuTriggerProps.onPress' was still be fired as expected. Though, it was not being fired on iOS. This update passes the 'menuTriggerProps.onPress' as the 'onClick' property of the trigger elements where needed.