feat: Remove vAlign property from TableStyles
Previous vertical alignment for all table styles was 'center' with the option to change it explicitly.
Only reason we ever changed it was to set it to "top" when using "rowHeight: flexible".
After revisiting the logic, it makes more sense to have the vAlign be set based on the `rowHeight` property, and not necessary to allow for the vAlign property.
Now 'rowHeight: fixed' defaults to 'center' vertical alignment, and 'rowHeight: flexible' defaults to 'top' vertical alignment.