feat: Add TableState.deleteRows
When introducing the 'TableState.keptSelectedRows' we assumed that the reason for a row being removed was due to filters. However, it is possible for a row to be removed intentionally due to deleting it. Adding a 'GridTableApi.deleteRows' method to ensure we do not keep these rows around in the 'keptSelectedRows' and actually delete them.