fix: SC-37802 Prevent DatePicker triggering virtual keyboard on touch devices (#906)
* fix: SC-37802 Prevent DatePicker triggering virtual keyboard on touch devices

Fixes issue where the DatePicker was often rendering underneath the virtual keyboard, preventing users from using the DatePicker.
To fix this, DateFieldBase now identifies if it is being used on a touch device. If it is, then it sets the new `TextFieldBase.preventEdit` property to set the input as "readonly". This forces touch screen users to only use the DatePicker for changing dates.

Additional changes include:
 - Move focus to the DatePicker once it is opened and return focus to input when DatePicker is closed.

Updates tests based on new logic for moving focus around

* PR Feedback

* Remove 'preventEdit' prop in favor of setting 'inputMode=none'. This prevents the virtual keyboard while still allowing the field to be editable