feat: Support 'fullWidth' prop on all text fields (#979)
Adds support for a `PresentationFieldProps.fullWidth` property. The
following components will respect the property:
- FormLines
- TextField
- TextAreaField
- RichTextField
- SelectField
- MultiSelectField
- DateField
- NumberField
- TreeSelectField

Updates the `FormLines` component to default to the `"lg"` size instead
of `"full"`. This will hopefully minimize the impact on existing layouts
as the previous default width for each field within `FormLines` was
550px. By using the `lg` size, we'll continue to default the fields
within `FormLines` to 550px.