feat: Update Switch component styling
Adds new 'labelStyle: "centered"'. This will stack the label on top of the switch and center both elements in the column.
Updates how tooltips are rendered for Switches. Now includes an icon to denote there is a tooltip, rather than requiring the user to discover it by hovering over it.

Adds this tooltip functionality to the Label component, though it is only currently used by Switch. At some point it would be good to move all Tooltips to this sort of UX.