fix: Form Labels to not extend full width of field
Kind of an annoying issue that I've see is you try to click away from a SelectField to close the menu, but instead the menu flickers and opens back up. This was due to clicking above the field, which happened to be still clicking on the label element which was taking up the full width of its container. When clicking a label element, then the focus is set to the associated field. By adding 'align-self: flex-start', then the label element no longer extends the full width of the container.