feat: Move to esm. (#1118)
This updates Beam to be a dual CJS/ESM package.

Ideally we'd just go to ESM-only, since vite is perfectly capable (and
prefers) pulling in ESM packages, even for downstream projects that are
technically still commonjs (i.e. internal-frontend), *but* until
downstream test suites also execute in ESM (*cough* Jest *cough*), we
still need a CommonJS set of artifacts as well.

In theory there are not any "*hard* breaking changes", but:

* Any `@homebound/beam/dist/...something...` import will now hard break
(and was not really supported before anyway) because we use `tsup` to
create bundled files
* An exception is `@homebound/beam/rtl` is now a blessed alternate entry
point/bundle that includes our testing utilities
* The `@homebound/beam/dist/utils/sb.tsx` import didn't really have
anything in it, so is no longer public
* Because `tsup` is doing bundling, it also does path rewriting, so we
can remove the `ts-patch` / `transform-paths` tsc plugin