fix: Attempting to fix ScrollShadow chromatic snapshots (#1136)
The idea behind this fix is to use data-chromatic="ignore", as defined
in https://www.chromatic.com/docs/ignoring-elements/. I'm thinking it'd
be better to always have these elements present to be ignored. Not sure
if even with the ignore attribute, having the shadow elements sometimes
render and sometimes not would cause the same inconsistencies (which is
the current issue).