feat: allow custom radio group name (#1155)
This PR adds support to provide a `name` prop to the `RadioGroupField`
component.

By default, the underlying <input type="radio"> element requires a name
attribute in order to be logically grouped with other radios. Without
name, each radio is treated as independent, meaning multiple radios can
be selected at the same time instead of behaving as a mutually exclusive
set.

This fixes some situations where the `RadioGroupField` is rendered
individually for a parent group.